feat(web): install martha and update navigation
