+++ web/src/scripts/app.js
@@ -1,7 +1,6 @@
 import { picoapp } from 'picoapp'
-import { size } from '@/util/dom'
+import { size } from 'martha'
 
-import anchors from '@/components/anchors'
 import burger from '@/components/burger'
 import lazyImage from '@/components/lazyImage'
 import projectSlider from '@/components/projectSlider'
@@ -9,7 +8,6 @@ import aboutCarousel from '@/components/aboutCarousel'
 import accordion from '@/components/accordion'
 
 const components = {
-  anchors,
   burger,
   lazyImage,
   projectSlider,
@@ -21,7 +19,6 @@ const state = {
   ...size(),
   isNavOpen: false,
   aboutCarouselIndex: 0,
-  percentScrolled: 0,
 }
 
 export default picoapp(components, state)
