+++ web/src/scripts/index.js
@@ -1,7 +1,7 @@
 import 'focus-visible'
 import app from '@/app'
 import router from '@/router'
-import { add, remove, on, size } from '@/util/dom'
+import { add, remove, on, size } from 'martha'
 import gsap from 'gsap'
 import loadFonts from '@/lib/font-loader'
 import { primaryInput } from 'detect-it'
