+++ web/src/templates/calls.liquid
@@ -7,7 +7,7 @@ landingPageHeader: false
 {% assign landingPageHeader = false %}
 
 
-<div id="calls" class="abs fill oy z1" data-router-view="home" data-component="anchors">
+<div id="calls" class="abs fill oy z1" data-router-view="home">
   <div>
     {% include header %}
       <section id="intro" class="rel ph20 m:ph40 l:ph80 pt90 m:pt100 pb20">
