+++ web/src/templates/includes/mobile-nav.liquid
@@ -2,22 +2,16 @@
   <nav class="ph25 pt80 rel z1">
     <ul class="mobile-nav__items">
       <li class="mobile-nav__item mb25">
-        <a class="c-tan f38 lh100 ls02em" href="#projects" data-router-disabled>Projects</a>
+        <a class="c-tan f38 lh100 ls02em" href="/projects">Projects</a>
       </li>
       <li class="mobile-nav__item mb25">
-        <a class="c-tan f38 lh100 ls02em" href="#about" data-router-disabled>About</a>
+        <a class="c-tan f38 lh100 ls02em" href="/people">People</a>
       </li>
       <li class="mobile-nav__item mb25">
-        <a class="c-tan f38 lh100 ls02em" href="#lab" data-router-disabled>Lab</a>
+        <a class="c-tan f38 lh100 ls02em" href="/events">Events</a>
       </li>
       <li class="mobile-nav__item mb25">
-        <a class="c-tan f38 lh100 ls02em" href="#people" data-router-disabled>People</a>
-      </li>
-      <li class="mobile-nav__item mb25">
-        <a class="c-tan f38 lh100 ls02em" href="#events" data-router-disabled>Events</a>
-      </li>
-      <li class="mobile-nav__item mb25">
-        <a class="c-tan f38 lh100 ls02em" href="#community" data-router-disabled>Community</a>
+        <a class="c-tan f38 lh100 ls02em" href="/news" >News</a>
       </li>
     </ul>
   </nav>
