+++ web/src/templates/includes/nav.liquid
@@ -1,38 +1,14 @@
 <nav class="nav df">
-  <div class="nav__column">
-    <div class="mb5">
-      <a href="#projects" class="nav__link dib sans f16 l:f18 pv5" data-router-disabled>
-        <span class="rel z1">Projects</span>
-      </a>
-    </div>
-    <div>
-      <a href="#about" class="nav__link dib sans f16 l:f18 pv5" data-router-disabled>
-        <span class="rel z1">About</span>
-      </a>
-    </div>
-  </div>
-  <div class="nav__column">
-    <div class="mb5">
-      <a href="#lab" class="nav__link dib sans f16 l:f18 pv5" data-router-disabled>
-        <span class="rel z1">Lab</span>
-      </a>
-    </div>
-    <div>
-      <a href="#people" class="nav__link dib sans f16 l:f18 pv5" data-router-disabled>
-        <span class="rel z1">People</span>
-      </a>
-    </div>
-  </div>
-  <div class="nav__column">
-    <div class="mb5">
-      <a href="#events" class="nav__link dib sans f16 l:f18 pv5" data-router-disabled>
-        <span class="rel z1">Events</span>
-      </a>
-    </div>
-    <div>
-      <a href="#community" class="nav__link dib sans f16 l:f18 pv5" data-router-disabled>
-        <span class="rel z1">Community</span>
-      </a>
-    </div>
-  </div>
+  <a href="/projects" class="nav__link dib sans f16 l:f18 pv5 mr50">
+    <span class="rel z1">Projects</span>
+  </a>
+  <a href="/people" class="nav__link dib sans f16 l:f18 pv5 mr50">
+    <span class="rel z1">People</span>
+  </a>
+  <a href="/events" class="nav__link dib sans f16 l:f18 pv5 mr50">
+    <span class="rel z1">Events</span>
+  </a>
+  <a href="/news" class="nav__link dib sans f16 l:f18 pv5 mr50">
+    <span class="rel z1">News</span>
+  </a>
 </nav>
