+++ web/src/templates/index.liquid
@@ -12,7 +12,7 @@ title: Home
 {% assign footer = landingPage.footer %}
 {% assign landingPageHeader = true %}
 
-<div class="landing abs fill oy z1" data-router-view="home" data-component="anchors">
+<div class="landing abs fill oy z1" data-router-view="home">
   <div>
     {% include header %}
     {% include landing-sections/intro %}
